+++ src/test/resources/org/sonar/plugins/delphi/standardLibrary/Vcl.Forms.pas
@@ -1,0 +1,13 @@
+unit Vcl.Forms;
+
+interface
+
+uses
+  System.Classes;
+
+type
+	TForm = class(TComponent)
+	end;
+implementation
+
+end.