+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `UnusedLocalVariablesRule` analysis rule, which flags unused local variables.
 - `UnusedFieldsRule` analysis rule, which flags unused fields.
 - `UnusedConstantsRule` analysis rule, which flags unused constants.
+- `UnusedGlobalVariablesRule` analysis rule, which flags unused global variables.
 - `UnusedTypesRule` analysis rule, which flags unused types.
 - `UnusedPropertiesRule` analysis rule, which flags unused properties.
 - `UnusedMethodsRule` analysis rule, which flags unused methods.